Secure initial development state
This project was initially made and need a repository to avoid loses and
prevent big regressions. The spend time was approximately one day.
Initial graphics consist of monochromatic squares were included, in the
future they will be change to better graphics.
Basic scripts to control the map, player and ice were made. The difficult
part is integrate game logic with the physical engine. A relational
pattern between the element nodes in each phase stage was defined.